move user view folder
