+++ client/src/components/FrontPage.tsx
@@ -9,7 +9,7 @@ import {NavigateFunction, useNavigate} from 'react-router-dom';
 import {UseQueryResult} from '@tanstack/react-query';
 
 import CourseTable from './front-page/CourseTable';
-import UsersView from './users-view/UsersView';
+import UsersView from './front-page/users-view/UsersView';
 
 import {useGetAllCourses, useGetCoursesOfUser} from '../hooks/useApi';
 import useAuth, {AuthContextType} from '../hooks/useAuth';
