+++ client/src/components/front-page/users-view/UsersView.tsx
@@ -6,7 +6,7 @@ import {JSX} from 'react';
 import {Box, Button, Typography} from '@mui/material';
 import {NavigateFunction, useNavigate} from 'react-router-dom';
 
-import useAuth, {AuthContextType} from '../../hooks/useAuth';
+import useAuth, {AuthContextType} from '../../../hooks/useAuth';
 import UsersTable from './UsersTable';
 
 export default function UsersView(): JSX.Element {
